social_network/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<3.13
gunicorn>=20.1,<20.2